CEDS-3438 Move CSE-EXRR validation to DeclarationHolder
