+++ app/controllers/declaration/DeclarationHolderChangeController.scala
@@ -24,6 +24,7 @@ import controllers.helpers.DeclarationHolderHelper._
 import controllers.helpers.MultipleItemsHelper
 import controllers.navigation.Navigator
 import forms.declaration.declarationHolder.DeclarationHolder
+import forms.declaration.declarationHolder.DeclarationHolder.DeclarationHolderFormGroupId
 import handlers.ErrorHandler
 import javax.inject.Inject
 import models.declaration.DeclarationHoldersData
