+++ app/views/declaration/declarationHolder/declaration_holder_edit_content.scala.html
@@ -15,9 +15,8 @@
  *@
 
 @import config.AppConfig
-@import controllers.helpers.DeclarationHolderHelper.DeclarationHolderFormGroupId
 @import forms.declaration.declarationHolder.DeclarationHolder
-@import forms.declaration.declarationHolder.DeclarationHolder.eoriSource
+@import forms.declaration.declarationHolder.DeclarationHolder.{DeclarationHolderFormGroupId, eoriSource}
 @import models.requests.JourneyRequest
 @import models.declaration.EoriSource._
 @import play.twirl.api.HtmlFormat
